refactor(master-v2): hoist double-play dashboard JSON mapper to core
Move snapshot_to_jsonable and display JSON v2 helpers from the WebUI route module into trading.master_v2.double_play_dashboard_display so route-independent tests do not depend on FastAPI. WebUI route delegates to the same callable; payload contract unchanged.
