+++ src/webui/double_play_dashboard_display_json_route_v0.py
@@ -10,9 +10,7 @@ See docs/ops/specs/MASTER_V2_DOUBLE_PLAY_WEBUI_READONLY_ROUTE_CONTRACT_V0.md.
 
 from __future__ import annotations
 
-from datetime import datetime, timezone
-from enum import Enum
-from typing import Any, Dict, Mapping
+from typing import Any, Dict
 
 from fastapi import APIRouter
 from fastapi.responses import JSONResponse
@@ -29,10 +27,9 @@ from trading.master_v2.double_play_composition import (
     compose_double_play_decision,
 )
 from trading.master_v2.double_play_dashboard_display import (
-    DashboardDisplayStatus,
     DoublePlayDashboardDisplaySnapshot,
-    DoublePlayDashboardPanel,
     build_dashboard_display_snapshot,
+    snapshot_to_jsonable,
 )
 from trading.master_v2.double_play_futures_input import (
     FuturesCandidateSnapshot,
@@ -78,36 +75,6 @@ router = APIRouter(
     prefix="/api/master-v2/double-play",
     tags=["master-v2", "double-play", "dashboard-display-readonly"],
 )
-
-_DISPLAY_JSON_LAYER_VERSION = "v2"
-
-_DISPLAY_PANEL_GROUP: Mapping[str, str] = {
-    "futures_input": "input",
-    "state_transition": "state",
-    "survival_envelope": "scope",
-    "strategy_suitability": "strategy",
-    "capital_slot_ratchet": "capital",
-    "capital_slot_release": "capital",
-    "composition": "composition",
-}
-
-_DISPLAY_SEVERITY_RANK: Mapping[DashboardDisplayStatus, int] = {
-    DashboardDisplayStatus.DISPLAY_READY: 0,
-    DashboardDisplayStatus.DISPLAY_WARNING: 10,
-    DashboardDisplayStatus.DISPLAY_MISSING: 20,
-    DashboardDisplayStatus.DISPLAY_BLOCKED: 30,
-    DashboardDisplayStatus.DISPLAY_ERROR: 40,
-}
-
-
-def _display_severity_rank(status: DashboardDisplayStatus) -> int:
-    return _DISPLAY_SEVERITY_RANK[status]
-
-
-def _assembled_at_iso_utc() -> str:
-    dt = datetime.now(timezone.utc).replace(microsecond=0)
-    return dt.isoformat().replace("+00:00", "Z")
-
 
 _GOOD_LIMITS = StaticHardLimits(
     max_notional=1.0,
@@ -379,48 +346,6 @@ def _build_static_double_play_dashboard_display_snapshot_v0() -> DoublePlayDashb
         capital_slot_release=rel,
         composition=comp,
     )
-
-
-def _jsonable_panel(panel: DoublePlayDashboardPanel, ordinal: int) -> Dict[str, Any]:
-    st = panel.status
-    status_val = st.value if isinstance(st, Enum) else str(st)
-    panel_group = _DISPLAY_PANEL_GROUP.get(panel.name, "unknown")
-    return {
-        "name": panel.name,
-        "status": status_val,
-        "summary": panel.summary,
-        "blockers": list(panel.blockers),
-        "missing_inputs": list(panel.missing_inputs),
-        "live_authorization": panel.live_authorization,
-        "is_authority": panel.is_authority,
-        "is_signal": panel.is_signal,
-        "ordinal": ordinal,
-        "panel_group": panel_group,
-        "severity_rank": _display_severity_rank(st),
-    }
-
-
-def snapshot_to_jsonable(snap: DoublePlayDashboardDisplaySnapshot) -> Dict[str, Any]:
-    """Convert display snapshot to JSON-serializable dict (enum values as strings)."""
-    overall = snap.overall_status
-    overall_val = overall.value if isinstance(overall, Enum) else str(overall)
-    return {
-        "display_layer_version": _DISPLAY_JSON_LAYER_VERSION,
-        "display_snapshot_meta": {
-            "source_kind": "static_display_v0",
-            "source_id": "webui_dashboard_display_static_v0",
-            "assembled_at_iso": _assembled_at_iso_utc(),
-        },
-        "panels": [_jsonable_panel(p, i) for i, p in enumerate(snap.panels)],
-        "overall_status": overall_val,
-        "no_live_banner_visible": snap.no_live_banner_visible,
-        "display_only": snap.display_only,
-        "trading_ready": snap.trading_ready,
-        "testnet_ready": snap.testnet_ready,
-        "live_ready": snap.live_ready,
-        "live_authorization": snap.live_authorization,
-        "warnings": list(snap.warnings),
-    }
 
 
 def build_static_dashboard_display_dict() -> Dict[str, Any]:
